Shift archive and settings commands to cli package
Update esak to v2
