+++ perdoo/cli/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["archive_app", "settings_app"]
+
+from perdoo.cli.archive import app as archive_app
+from perdoo.cli.settings import app as settings_app
