+++ pyproject.toml
@@ -35,7 +35,7 @@ classifiers = [
 ]
 dependencies = [
   "comicfn2dict >= 0.2.4",
-  "esak >= 1.3.2",
+  "esak >= 2.0.0",
   "himon >= 0.6.1",
   "lxml >= 5.3.0",
   "mokkari >= 3.4.0",
